refactor(rulegen): move proxy dectection to a separated file
